Add links to source code in ExDoc
This adds links from ExDoc to the source code on Github, so it's easy for
developers to see the implementation of a module/function. This also sets the
main page for the docs to the `Toml` module, instead of the current "API
Reference" page. This configuration seems to make sense because the `Toml`
module contains the Readme as its `@moduledoc` and is the place where most
people would go to get started with this library.