Fix now incorrect assumption in test suite
`Toml.Test.JsonConverter` made an assumption that a list which started
with a map was to be treated as a TOML table array.  The new mixed-type
array tests specifically include a test for this assumption because, as
the note in the test points out, the assumption was reasonable prior to
TOML v1.0 but not after.

Remove the table array `to_typed_value` function head targeting table
arrays.  It looks like this was function head was designed to ensure
that the now obsolete assumption held true; table arrays will still be
parsed correctly without it and the now allowed mixed arrays as well.